+++ src/aioway/fate/fate.py
@@ -8,10 +8,10 @@ import typing
 
 from torch import _ops
 
-from aioway._common._types import dcls_no_repr
-from aioway._common.decomps import find_nested_tensors
-from aioway._common.renders import render_fcall
 from aioway._keyed import Keyed
+from aioway._types import dcls_no_repr
+from aioway.decomps import find_nested_tensors
+from aioway.renders import render_fcall
 
 __all__ = ["Fate", "find_fate", "all_fates"]
