+++ src/aioway/fn/_on_off.py
@@ -8,7 +8,7 @@ import dataclasses as dcls
 import logging
 import typing
 
-from aioway._common._types import Stack
+from aioway._types import Stack
 
 __all__ = ["OnOffCtx", "OnOffStack"]
