+++ src/aioway/io/videos.py
@@ -8,7 +8,7 @@ import typing
 import torch
 from torchcodec import decoders as dec
 
-from aioway._common._platforms import num_threads
+from aioway._platforms import num_threads
 from aioway.fake import enabled_fake_mode, torch_set_fake_mode_func
 from aioway.schemas import Attr
 from aioway.tags import IsVideoTag
