+++ src/aioway/mess/init/dense.py
@@ -2,7 +2,7 @@
 
 from torch import nn
 
-from aioway._common._types import dcls_no_repr
+from aioway._types import dcls_no_repr
 
 from .init import MessInit
