+++ src/aioway/mess/init/losses.py
@@ -4,7 +4,7 @@ import typing
 
 from torch import nn
 
-from aioway._common._types import dcls_no_repr
+from aioway._types import dcls_no_repr
 
 from .init import MessInit
