+++ src/aioway/schemas/sets.py
@@ -12,7 +12,7 @@ import numpy as np
 import tensordict as td
 import torch
 
-from aioway._common._typing import is_dict_of_str_to, is_list_of
+from aioway._typing import is_dict_of_str_to, is_list_of
 from aioway.schemas import LayoutLike
 
 from .attrs import Attr, attr
