+++ tests/datasets/streams/test_maps.py
@@ -7,7 +7,7 @@ from collections import abc as cabc
 import pytest
 import tensordict as td
 
-from aioway._common._torch import tdict_all_equal, tdict_rename
+from aioway._torch import tdict_all_equal, tdict_rename
 from aioway.datasets import (
     ApplyStream,
     CacheStream,
