+++ tests/might/test_might.py
@@ -5,8 +5,8 @@ import typing
 import pytest
 from torch import nn
 
-from aioway._common.renders import render_fcall
 from aioway.mess import MessInit, Sequential, find_might
+from aioway.renders import render_fcall
 
 
 class _ModuleOpts(typing.NamedTuple):
